Initial patch for USB tracking using inotify
Currentlty, USB connect/disconnect events are triggered using usb_moded.
But usb_moded changes are not ready and also, this mechanism can avoid a
dependency on usb_moded. This mechanism would use inotify to keep track
of the devices created (/dev/ttyGS1) and appropriately signal the
plugins with the connectivityChanged() signal

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>