Created OOPClientPlugin and OOPServerPlugin implementation templates
The OOPClientPlugin and OOPServerPlugin instances using dbus to interact
with the process plugins that have been started by PluginManager. All
the signals/slots that are triggered because of the plugins or the
framework are handled by these classes and they inturn
marshall the messages and send them to the process plugins. Similarly,
for the messages received from the process plugins, these classes
unmarshall them and send the appropriate signals

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>