Using Qt methods to log output to stdout and removed logging to file
The Qt debug methods, qDebug(), qFatal() etc. are used to log the output
to stdout rather than to a file (using the Buteo message handler). This
choice is valid, since in production devices, it is not expected that
msyncd writes to files (which it does not in the prev. way too). In dev
environment, developers can start msyncd from cmdline and check the
output there itself, or redirect the ouput to a file. msyncd now does
not initialize the Logger::instance()

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>