Completed plugin_main.cpp and also the outer shell for plugins
The plugin_main.cpp, PluginServiceObj.h and PluginServiceObj.cpp files
form a shell around a plugin to perform the communication between msyncd
and the plugin process. The only change that needs to be done in the
plugin code is to add an extra config for generating the exe and to
include the above files in the make. The syncmlclient plugin code has
been modified to prove this mechanism

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>