Patch to create the callback class towards msyncd from plugins
This patch enables the callback mechanism from plugins to msyncd. The
daemon has information related to scheduling, storage providers,
connectivity which can be used by the plugins to perform sync actions.
Also increased the sleep in PluginManager for process state to 1sec,
since 200msec does not seem to work. Also added method
isConnectivityAvailable to the dbus methods in msyncd

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>